New name proposal for Stepper
In the column other XAML it was written the name NumericUpDown, but WPF has never had such a control. There was a NUD in the Silverlight toolkit as well as in Windows Forms, but the former is obsolete and the later is not a XAML framework. On the other hand, recently the WinUI team has released a NumberBox control with a very similar API than the NumericUpDown. So I think, NumberBox would be name the align the most. Optionally, I wrote the name NumericUpDown since that is the classical one, so to speak.